pyproject.toml
[project]
name = "csa-copilot"
version = "3.0.0"
description = "CSA-Copilot: AI-powered presentation and demo builder for Microsoft Cloud Solution Architects"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "github-copilot-sdk==0.1.29",
    "mcp>=1.0",
    "pydantic>=2.0",
    "prompt-toolkit>=3.0.43",
    "rich>=13.0",
    "python-pptx>=1.0.2",
    "markitdown>=0.1.0",
    "pyyaml>=6.0",
    # Desktop server layer
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "websockets>=13.0",
    "pillow>=11.0",
]

[project.scripts]
csa-copilot = "app:main_entry"

[tool.setuptools]
py-modules = ["app", "router", "tools", "ui", "pricing", "store", "collector", "queries"]
packages = ["agents", "commands"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["."]
omit = [
    "app.py",
    "ui.py",
    "tests/*",
    "skills/*",
    "agent_defs/*",
    "scripts/*",
    "outputs/*",
    "plans/*",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "if TYPE_CHECKING",
]
